in tests do not remove docker container if on circleci

+++ tests/test.sh
@@ -18,7 +18,9 @@ cd ..
 echo "Running unit tests..."
 python3 test.py
 
-# Remove DB container
-echo "Removing DB container..."
-docker kill ${db_docker_id}
-docker rm -f ${db_docker_id}
+# Remove DB container (if not on CircleCI)
+if [ "$CIRCLECI" = false ] ; then
+    echo "Removing DB container..."
+    docker kill ${db_docker_id}
+    docker rm -f ${db_docker_id}
+fi
